requirements.txt
streamlit==1.37.1
pdfplumber>=0.7.1
pdfminer.six>=20201018
pytesseract>=0.3.10
pillow>=9.0.0
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.2
langdetect>=1.0.9
scikit-learn>=1.2.2
numpy<2.0  # IMPORTANT: scikit-learn & some wheels may still require numpy < 2 for your env
